move logic classes to logic package
