+++ src/main/kotlin/logic/SearchAlgorithm.kt
@@ -1,4 +1,4 @@
-package utils
+package logic
 
 interface SearchAlgorithm {
     fun search(fullText: String, pattern: String): Boolean
